Split constants into constants and globals
